requirements.txt
streamlit
transformers
torch
diffusers
Pillow
numpy